// src/common/enums/file-upload.enum.ts
export enum EFileType {
  IMAGE = 'IMAGE',
  VIDEO = 'VIDEO',
  WORD = 'WORD',
  EXCEL = 'EXCEL',
  PDF = 'PDF',
  POWERPOINT = 'POWERPOINT',
  TEXT = 'TEXT',
  ARCHIVE = 'ARCHIVE',
  OTHER = 'OTHER',
}

export enum EFileStatus {
  ACTIVATED = 'ACTIVATED',
  DEACTIVATED = 'DEACTIVATED',
}

export enum EFileEntity {
  USER = 'USER',
  PRODUCT = 'PRODUCT',
  VARIANT = 'VARIANT',
  ORDER = 'ORDER',
  OTHER = 'OTHER',
}

export enum EFileClient {
  WEBSITE = 'website',
  ADMIN = 'admin',
}